+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.1.1] - Add custom global shortcuts.
+
+* You can now config shortcuts that will be allow in all pages were globalShortcuts is true
+
 ## [0.1.0] - Release package.
 
 * Release first version of the widget